requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-ckeditor==5.9.0
django-countries==6.1.2
django-filter==2.2.0
django-import-export==2.0.2
django-js-asset==1.2.2
django-localflavor==3.0.1
djangorestframework==3.11.0
et-xmlfile==1.0.1
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
promise==2.3
python-dateutil==2.8.1
python-stdnum==1.13
pytz==2019.3
PyYAML==5.3.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
xlrd==1.2.0
xlwt==1.3.0
